commit for trigger test 3

+++ src/main/java/com/example/calculator/CalculatorApplication.java
@@ -8,7 +8,7 @@ import	org.springframework.stereotype.Service;
 @Service
 public class CalculatorApplication {
 
-	long sum((int a, int b){
+	int sum((int a, int b){
 		return (a+b);
 	}
